Downcase UTF-8 file encoding name
Python and Emacs (and others?) recognize a special directive line in a
file that identifies what encoding the file is encoded in. See Python
PEP 263. For example:

The general form of the directive is:

where xxx is the name of a codec. Python codec names are lower case
with underscores used to seperate words.

In both Python and Emacs one can create aliases for the codecs so you
can use an alternate name to refer to the same codec.

Python is forgiving with respect to case, underscore and
hyphens. Python will automatically create an alias for a codec name by
downcasing it and replacing hyphens with underscores, thus "UTF-8" is
actually an alias for the "utf_8" codec. Unfortunately emacs does not
automatically create such aliases, although one can add aliases via a
custom initialization file, but doing so requires every user using
emacs to edit the files to manually create their own aliases.

If you try to write a file in emacs with the "UTF-8" codec name it
won't recognize it as "utf-8", instead you'll get errors like this:

  Warning (mule): Invalid coding system ‘UTF-8’ is specified
  for the current buffer/file by the :coding tag.
  It is highly recommended to fix it before writing to a file.

and you must force the file to be written by responding to additional
propmpts.

This patch simply downcases the the "UTF-8" codec name to "utf-8" so
that both Python and Emacs will accept the codec name.

License: MIT
Signed-off-by: John Dennis <jdennis@redhat.com>